requirements.txt
flask==3.0.0
efficientnet==1.1.1
tensorflow==2.15.0
numpy==1.26.0
keras==2.15.0
Pillow==10.1.0
flask_cors==4.0.0